Adding _setenv and _unsetenv functions

+++ _strlen.c
@@ -4,9 +4,9 @@
  * @s: pointer to our char
  * Return: length of a string
  */
-int _strlen(char *s)
+size_t _strlen(char *s)
 {
-	int a = 0;
+	size_t a = 0;
 
 	while (*(s + a) != '\0')
 		a++;
